finished trunc poisson support, added test    added run option to admbControl    don't erase .std file, just warn
git-svn-id: svn://svn.r-forge.r-project.org/svnroot/glmmadmb@151 edb9625f-4e0d-4859-8d74-9fd3b1da38cb

+++ pkg/R/glmm.admb.R
@@ -1,13 +1,15 @@
 pkgname <- "glmmADMB"
 
 admbControl <- function(impSamp=0,
-                         maxfn=500,
-                         imaxfn=500,
-                         noinit=TRUE,
-                         shess=TRUE,
-                         ZI_kluge=FALSE) {
+                        maxfn=500,
+                        imaxfn=500,
+                        noinit=TRUE,
+                        shess=TRUE,
+                        run=TRUE,
+                        ZI_kluge=FALSE) {
   list(impSamp=impSamp,maxfn=maxfn,imaxfn=imaxfn,noinit=noinit,shess=shess,
-       ZI_kluge=ZI_kluge)
+       ZI_kluge=ZI_kluge,run=run)
+  ## FIXME: do something clever with formals/match.call() ?
 }
   
 mcmcControl <- function(mcmc=1000,
@@ -89,6 +91,7 @@ glmmadmb <- function(formula, data, family="poisson", link,start,
   impSamp <- admb.opts$impSamp
   maxfn <- admb.opts$maxfn
   imaxfn <- admb.opts$imaxfn
+  run <- admb.opts$run
   
   if (use_tmp_dir <- missing(save.dir)) {
     repeat {
@@ -136,7 +139,7 @@ glmmadmb <- function(formula, data, family="poisson", link,start,
     family <- "nbinom"
   }
   if (family=="gaussian") stop("gaussian family not yet debugged")
-  if (family=="truncpoiss") stop("truncated Poisson support is EXPERIMENTAL/may not be supported other than on linux-32")
+  if (family=="truncpoiss") warning("truncated Poisson support is EXPERIMENTAL/may not be supported other than on linux-32")
     
   like_type_flag <- switch(tolower(family),poisson=0,binomial=1,nbinom=2,gamma=3,beta=4,gaussian=5,
                            truncpoiss=6,
@@ -145,7 +148,7 @@ glmmadmb <- function(formula, data, family="poisson", link,start,
 
 
   if (missing(link)) {
-    link <- switch(family, binomial=, beta="logit", nbinom=, poisson=, gamma="log",gaussian="identity")
+    link <- switch(family, binomial=, beta="logit", nbinom=, poisson=, truncpoiss=, gamma="log",gaussian="identity")
   }
   linkfun <- switch(link,log=log,logit=qlogis,probit=qnorm,inverse=function(x) {1/x},
                     cloglog=function(x) {log(-log(1-x))},
@@ -312,25 +315,26 @@ glmmadmb <- function(formula, data, family="poisson", link,start,
   dat_write(file_name, dat_list)
   pin_write(file_name, pin_list)
   std_file <- paste(file_name, ".std", sep="")
-  if(file.exists(std_file))
-    file.remove(std_file)
+  if(file.exists(std_file) && run) warning("file",std_file,"exists: overwriting")
+  ## file.remove(std_file)
 
+  if (run) {
   ## FIXME: for what platforms do we really need to copy the binary?
   ##  can't we just run it in place?  Or does it do something silly and produce
   ##  output in the directory in which the binary lives rather than the
   ##  current working directory (feel like I struggled with this earlier
   ##  but have now forgotten -- ADMB mailing list archives??)
-  if (platform=="windows") {
-    cmd <- paste("\"",bin_loc, "\"", " ", cmdoptions, sep="")
-    shell(cmd, invisible=TRUE)
-  } else  {
-    file.copy(bin_loc,".")
-    Sys.chmod(file_name,mode="0755") ## file.copy strips executable permissions????
-    cmd2 <- paste("./", file_name, " ", cmdoptions, sep="")
-    sys.result <- system(cmd2,intern=!verbose)
-    unlink(file_name)
+    if (platform=="windows") {
+      cmd <- paste("\"",bin_loc, "\"", " ", cmdoptions, sep="")
+      shell(cmd, invisible=TRUE)
+    } else  {
+      file.copy(bin_loc,".")
+      Sys.chmod(file_name,mode="0755") ## file.copy strips executable permissions????
+      cmd2 <- paste("./", file_name, " ", cmdoptions, sep="")
+      sys.result <- system(cmd2,intern=!verbose)
+      unlink(file_name)
+    }
   }
-
   if(!file.exists(std_file))
     stop("The function maximizer failed")
   tmp <- read.table(paste(file_name,"std",sep="."), skip=1,as.is=TRUE)
